# perf: fix lazy loading of core rules