--- a/tests/fixtures/lazy-loading-rules/.eslintrc.js
+++ b/tests/fixtures/lazy-loading-rules/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+    root: true,
+    rules: {
+        semi: 2
+    }
+};