--- a/tests/fixtures/lazy-loading-rules/foo.js
+++ b/tests/fixtures/lazy-loading-rules/foo.js
@@ -1,0 +1,1 @@
+/* content is not necessary */